requirements.txt
requests
beautifulsoup4
lxml
tldextract
python-dateutil
html2text
